♻️ refactor pydantic stuff into its own file
