+++ tests/runtime/test_http_server.py
@@ -17,8 +17,7 @@ Tests for the caikit HTTP server
 # Standard
 from contextlib import contextmanager
 from dataclasses import dataclass
-from typing import Dict, List, Union, get_args
-import enum
+from typing import Dict
 import json
 import os
 import signal
@@ -27,23 +26,13 @@ import tempfile
 # Third Party
 from fastapi.testclient import TestClient
 import numpy as np
-import pydantic
 import pytest
 import requests
 import tls_test_tools
 
-# First Party
-from py_to_proto.dataclass_to_proto import Annotated
-
 # Local
-from caikit.core import MODEL_MANAGER, DataObjectBase, dataobject
-from caikit.core.data_model.base import DataBase
-from caikit.interfaces.common.data_model.stream_sources import File
-from caikit.interfaces.nlp.data_model import GeneratedTextStreamResult, GeneratedToken
+from caikit.core import MODEL_MANAGER
 from caikit.runtime import http_server
-from caikit.runtime.service_generation.data_stream_source import make_data_stream_source
-from sample_lib.data_model import SampleInputType, SampleOutputType
-from sample_lib.data_model.sample import SampleTrainingType
 from tests.conftest import temp_config
 from tests.runtime.conftest import (
     ModuleSubproc,
@@ -249,235 +238,6 @@ def test_services_disabled(open_port, enabled_services):
             # assert (server.global_train_servicer and enable_training) or (
             #     server.global_train_servicer is None and not enable_training
             # )
-
-
-## Functional Tests #######################################################################
-
-
-def test_build_dm_object_simple(runtime_http_server):
-    """Test building our simple DM objects through pydantic objects"""
-
-    # get our DM class
-    sample_input_dm_class = DataBase.get_class_for_name("SampleInputType")
-    # get pydantic model for our DM class
-    sample_input_pydantic_model = http_server.PYDANTIC_TO_DM_MAPPING.get(
-        sample_input_dm_class
-    )
-    # build our DM object using a pydantic object
-    sample_input_dm_obj = runtime_http_server._build_dm_object(
-        sample_input_pydantic_model(name="Hello world")
-    )
-
-    # assert it's our DM object, all fine and dandy
-    assert isinstance(sample_input_dm_obj, DataBase)
-    assert sample_input_dm_obj.to_json() == '{"name": "Hello world"}'
-
-
-def test_build_dm_object_datastream_jsondata(runtime_http_server):
-    """Test building our datastream DM objects through pydantic objects"""
-
-    # get our DM class
-    datastream_dm_class = DataBase.get_class_for_name(
-        "DataStreamSourceSampleTrainingType"
-    )
-    # get pydantic model for our DM class
-    datastream_pydantic_model = http_server.PYDANTIC_TO_DM_MAPPING.get(
-        datastream_dm_class
-    )
-    # build our DM Datastream JsonData object using a pydantic object
-    datastream_dm_obj = runtime_http_server._build_dm_object(
-        datastream_pydantic_model(data_stream={"data": [{"number": 1}, {"number": 2}]})
-    )
-
-    # assert it's our DM object, all fine and dandy
-    assert isinstance(datastream_dm_obj, DataBase)
-    assert (
-        datastream_dm_obj.to_json()
-        == '{"jsondata": {"data": [{"number": 1}, {"number": 2}]}}'
-    )
-
-
-def test_build_dm_object_datastream_file(runtime_http_server):
-    # get our DM class
-    datastream_dm_class = DataBase.get_class_for_name(
-        "DataStreamSourceSampleTrainingType"
-    )
-    # get pydantic model for our DM class
-    datastream_pydantic_model = http_server.PYDANTIC_TO_DM_MAPPING.get(
-        datastream_dm_class
-    )
-
-    # build our DM Datastream File object using a pydantic object
-    datastream_dm_obj = runtime_http_server._build_dm_object(
-        datastream_pydantic_model(data_stream={"filename": "hello"})
-    )
-
-    # assert it's our DM object, all fine and dandy
-    assert isinstance(datastream_dm_obj, DataBase)
-    assert isinstance(datastream_dm_obj.data_stream, File)
-    assert datastream_dm_obj.to_json() == '{"file": {"filename": "hello"}}'
-
-
-@pytest.mark.parametrize(
-    "input, output",
-    [
-        (np.integer, int),
-        (np.floating, float),
-        (int, int),
-        (float, float),
-        (bool, bool),
-        (str, str),
-        (bytes, bytes),
-        (type(None), type(None)),
-        (enum.Enum, enum.Enum),
-        (Annotated[str, "blah"], str),
-        (Union[str, int], Union[str, int]),
-        (List[str], List[str]),
-        (List[Annotated[str, "blah"]], List[str]),
-        (Dict[str, int], Dict[str, int]),
-        (Dict[Annotated[str, "blah"], int], Dict[str, int]),
-    ],
-)
-def test_get_pydantic_type(input, output):
-    assert http_server.RuntimeHTTPServer._get_pydantic_type(input) == output
-
-
-def test_get_pydantic_type_union():
-    union_type = Union[SampleInputType, SampleOutputType]
-    return_type = http_server.RuntimeHTTPServer._get_pydantic_type(union_type)
-    assert all(
-        issubclass(ret_type, pydantic.BaseModel) for ret_type in get_args(return_type)
-    )
-
-
-def test_get_pydantic_type_DM():
-    # DM case
-    sample_input_dm_class = DataBase.get_class_for_name("SampleInputType")
-    sample_input_pydantic_model = http_server.RuntimeHTTPServer._get_pydantic_type(
-        sample_input_dm_class
-    )
-
-    assert issubclass(sample_input_pydantic_model, pydantic.BaseModel)
-    assert sample_input_pydantic_model in http_server.PYDANTIC_TO_DM_MAPPING
-    assert sample_input_pydantic_model is http_server.PYDANTIC_TO_DM_MAPPING.get(
-        sample_input_dm_class
-    )
-
-
-def test_get_pydantic_type_throws_random_type():
-    # error case
-    with pytest.raises(TypeError):
-        http_server.RuntimeHTTPServer._get_pydantic_type("some_random_type")
-
-
-def test_pydantic_wrapping_with_enums():
-    """Check that the pydantic wrapping works on our data models when they have enums"""
-    # The NLP GeneratedTextStreamResult data model contains enums
-
-    # Check that our data model is fine and dandy
-    token = GeneratedToken(text="foo")
-    assert token.text == "foo"
-
-    # Wrap the containing data model in pydantic
-    http_server.RuntimeHTTPServer._dataobject_to_pydantic(GeneratedTextStreamResult)
-
-    # Check that our data model is _still_ fine and dandy
-    token = GeneratedToken(text="foo")
-    assert token.text == "foo"
-
-
-def test_pydantic_wrapping_with_lists(runtime_http_server):
-    """Check that pydantic wrapping works on data models with lists"""
-
-    @dataobject(package="http")
-    class BarTest(DataObjectBase):
-        baz: int
-
-    @dataobject(package="http")
-    class FooTest(DataObjectBase):
-        bars: List[BarTest]
-
-    foo = FooTest(bars=[BarTest(1)])
-    assert foo.bars[0].baz == 1
-
-    runtime_http_server._dataobject_to_pydantic(FooTest)
-
-    foo = FooTest(bars=[BarTest(1)])
-    assert foo.bars[0].baz == 1
-
-
-def test_dataobject_to_pydantic_simple_DM():
-    """Test that we can create a pydantic model from a simple DM"""
-    sample_input_dm_class = DataBase.get_class_for_name("SampleInputType")
-    sample_input_pydantic_model = http_server.RuntimeHTTPServer._dataobject_to_pydantic(
-        sample_input_dm_class
-    )
-    model_instance = sample_input_pydantic_model.model_validate_json(
-        '{"name": "world"}'
-    )
-    assert {"name": str} == sample_input_pydantic_model.__annotations__
-    assert issubclass(sample_input_pydantic_model, pydantic.BaseModel)
-    assert sample_input_pydantic_model in http_server.PYDANTIC_TO_DM_MAPPING
-    assert model_instance.name == "world"
-
-
-def test_dataobject_to_pydantic_simple_DM_extra_forbidden_throws():
-    """Test that if we forbid extra values, then we raise if we pass in extra values"""
-    sample_input_dm_class = DataBase.get_class_for_name("SampleInputType")
-
-    sample_input_pydantic_model = http_server.RuntimeHTTPServer._dataobject_to_pydantic(
-        sample_input_dm_class
-    )
-    # sample_input_pydantic_model_extra_forbidden doesn't allow anything extra
-    with pytest.raises(pydantic.ValidationError) as e1:
-        sample_input_pydantic_model.model_validate_json('{"blah": "world"}')
-    assert "Extra inputs are not permitted" in e1.value.errors()[0]["msg"]
-
-    with pytest.raises(pydantic.ValidationError) as e2:
-        sample_input_pydantic_model.model_validate_json(
-            '{"name": "world", "blah": "world"}'
-        )
-    assert "Extra inputs are not permitted" in e2.value.errors()[0]["msg"]
-
-
-def test_dataobject_to_pydantic_oneof():
-    """Test that we can create a pydantic model from a DM with a Union"""
-    make_data_stream_source(SampleTrainingType)
-    sample_input_dm_class = DataBase.get_class_for_name(
-        "DataStreamSourceSampleTrainingType"
-    )
-    data_stream_source_pydantic_model = (
-        http_server.RuntimeHTTPServer._dataobject_to_pydantic(sample_input_dm_class)
-    )
-
-    assert issubclass(data_stream_source_pydantic_model, pydantic.BaseModel)
-    assert {
-        "data_stream": Union[
-            http_server.PYDANTIC_TO_DM_MAPPING.get(sample_input_dm_class.JsonData),
-            http_server.PYDANTIC_TO_DM_MAPPING.get(sample_input_dm_class.File),
-            http_server.PYDANTIC_TO_DM_MAPPING.get(sample_input_dm_class.ListOfFiles),
-            http_server.PYDANTIC_TO_DM_MAPPING.get(sample_input_dm_class.Directory),
-            http_server.PYDANTIC_TO_DM_MAPPING.get(sample_input_dm_class.S3Files),
-        ]
-    } == data_stream_source_pydantic_model.__annotations__
-    assert data_stream_source_pydantic_model in http_server.PYDANTIC_TO_DM_MAPPING
-
-    assert issubclass(
-        http_server.PYDANTIC_TO_DM_MAPPING.get(sample_input_dm_class.JsonData),
-        type(
-            data_stream_source_pydantic_model.model_validate_json(
-                '{"data_stream": {"data": [{"number": 1}]}}'
-            ).data_stream
-        ),
-    )
-    assert issubclass(
-        http_server.PYDANTIC_TO_DM_MAPPING.get(sample_input_dm_class.File),
-        type(
-            data_stream_source_pydantic_model.model_validate_json(
-                '{"data_stream": {"filename": "file1"}}'
-            ).data_stream
-        ),
-    )
 
 
 ## Inference Tests #######################################################################
